+++ Diminuto/tst/unittest-buffer.c
@@ -48,7 +48,7 @@ int main(int argc, char * argv[])
         ASSERT(sizeof(overhead.align) == 0);
         ASSERT(sizeof(overhead.payload) == 0);
         ASSERT(sizeof(overhead.header.item) == sizeof(overhead.header.next));
-        CHECKPOINT("alignment: %d", (((uintptr_t)(&overhead)) & 0x7));
+        CHECKPOINT("alignment: %llu", (long long unsigned int)(((uintptr_t)(&overhead)) & 0x7));
         ASSERT((((uintptr_t)(&overhead)) & 0x7) == 0);
         ASSERT((((uintptr_t)(&(overhead.align))) & 0x7) == 0);
         ASSERT((((uintptr_t)(&(overhead.payload))) & 0x7) == 0);
